flow-api/Jobs_Uduma_ReportUpdate/requirements.txt
pytest==5.0.1
Hypothesis==3.58
PyMySQL==0.9.3
psycopg2==2.7.7
psycopg2-binary==2.8.4
SQLAlchemy==1.3.13
sqlalchemy-migrate==0.13.0
SQLAlchemy-Utils==0.34.0
GeoAlchemy2==0.6.3
requests==2.21.0
numpy==1.14.2
pandas==0.25.0
